+++ vitest.config.js
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+    test: {
+        globals: true,
+        environment: 'node',
+        setupFiles: ['./vitest.setup.js'],
+        exclude: [
+            '**/node_modules/**',
+            '**/web/**'
+        ]
+    }
+});
